Enhance flash driver of STM32

+++ bsp_stm32f4/flash/flash_drv.h
@@ -36,7 +36,7 @@
 
 bool flash_init(void);
 bool flash_errase(uint32_t addr, uint32_t len);
-bool flash_wr(uint32_t addr, uint32_t* array, uint32_t array_len);
+bool flash_wr(uint32_t addr, uint8_t* array, uint32_t array_len);
 bool flash_read(uint32_t in_flash_addr, uint8_t* rx_array, uint32_t array_len);
 bool is_errased(uint32_t addr, uint32_t size);
 bool is_flash_spare(uint32_t flash_addr, uint32_t size);
